requirements.txt
requests
packaging
python-dateutil
